requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
opendartreader>=0.2.0
python-dotenv>=1.0.0
